% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the LocAlignR Shiny app}
\usage{
run_app(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, open the app in a browser.}

\item{...}{Passed to shiny::runApp().}
}
\value{
None. Starts a Shiny app.
}
\description{
Run the LocAlignR Shiny app
}
